LayoutTests/imported/w3c/web-platform-tests/resource-timing/SO-XO-SO-redirect-chain-tao.https-expected.txt


PASS Verify that cross origin resources' timings are not exposed when same-origin=>cross-origin=>same-origin redirects have no `Timing-Allow-Origin:` headers.
FAIL Verify that cross origin resources' timings are not exposed when same-origin=>cross-origin=>same-origin redirects have `Timing-Allow-Origin:` headers only on some of the responses. assert_equals: redirectStart should be 0 expected 0 but got 136
PASS Verify that cross origin resources' timings are exposed when cross-origin redirects have `Timing-Allow-Origin: *` headers
FAIL Verify that cross origin resources' timings are not exposed when same-origin=>cross-origin=>same-origin redirects have `Timing-Allow-Origin:` headers with a specific origin. assert_equals: redirectStart should be 0 expected 0 but got 283.00000000000006
